+++ src/index.ts
@@ -1,4 +1,4 @@
-import FragmentPlugin from './inc/FragmentPlugin.js';
-import type { Rule, Route } from './inc/FragmentPlugin.js';
+import FragmentPlugin from './SwupFragmentPlugin.js';
+import type { Rule, Route } from './SwupFragmentPlugin.js';
 export default FragmentPlugin;
 export { Rule as FragmentRule, Route as FragmentRoute };
